Refactor ChatBuble and ChatCard widgets
This commit refactors the ChatBuble and ChatCard widgets by removing unnecessary code and comments. The GroupChatBubleFriend widget is now a stateless widget since it doesn't require any state. The commented out code in ChatCard has been removed and a comment has been added to explain the difference between using context.read and BlocProvider.of to access the ChatCubit instance.