Add Firebase Cloud Messaging (FCM) push notifications functionality
This commit adds the ability to send push notifications using Firebase Cloud Messaging (FCM) to the app. It includes a new class `AccessTokenFirebase` that retrieves an access token for FCM, and a new method `sendNotification` in the `FirebaseHelper` class that sends a notification to a specific user using their push token. The `ChatCubit` and `FireStorage` classes have been updated to call the `sendNotification` method when a new message or image is sent. The `UserDataCubit` class has been updated to store the user model as a property, and the `SettingsView` has been updated to display the user's name and profile picture. Additionally, the `googleapis_auth` and `google_identity_services_web` packages have been added as dependencies.