Refactor login process and add getUserData call
In this commit, the login process in the AuthCubit has been refactored to include an emit call for a successful login. Additionally, in the LoginView, after a successful login, a call to getUserData has been added to the UserDataCubit. The constructor of the UserDataCubit has been modified to remove the initial call to getUserData.