Refactor code and clear push token on sign out
- Import `FirebaseMessaging` to clear push token on sign out
- Refactor `signOutDialog` method to include a `BuildContext` parameter
- Clear push token on sign out by calling `FirebaseMessaging.instance.deleteToken()`
- Remove commented out code in `UserDataCubit` class

The changes in this commit refactor the code and add functionality to clear the push token on sign out. The `signOutDialog` method is refactored to include a `BuildContext` parameter for better code organization. The push token is cleared on sign out by calling `FirebaseMessaging.instance.deleteToken()`. Additionally, commented out code in the `UserDataCubit` class is removed.